Add query methods for Statement which take ownership
This is in the same spirit as `Connection::query()`, which also takes ownership
of the statement it used.
An actual use case of this can be wanting to store the ResultSet in a struct for
later usage, while at the same time having the ability to use StatementBuilder
to set parameters.

Signed-off-by: Christoph Heiss <contact@christoph-heiss.at>