add build matrix to GitHub Action (OS, rustc version)
this allows testing the build also on MacOS and Windows (so far only
linux was tested) and testing against the minimum supported rustc
version (which should always be aligned with what's documented in the
README).

all other tests are still only run on linux to avoid unnecessary
overhead during execution (and it might be that running Oracle on those
systems might require more work).